LayoutTests/platform/chromium/accessibility/listbox-hit-test-expected.txt


This tests that hit testing works on a listbox.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.role is 'AXRole: AXStaticText'
PASS element.stringValue is 'AXValue: test option that spans the width of the cell. test option that spans the width of the cell'
PASS element.parentElement().role is 'AXRole: AXList'
PASS successfullyParsed is true

TEST COMPLETE
